Add simple CLI program for testing
Now upon `zig build`, a program called `basedirs` will be built, which
simply dumps all basedirs to stdout as YAML

I've also removed `exePath` as it's apparently in Zig's stdlib as
`std.io.selfExePath` and made a fallback for POSIX systems that obtains
the home directory from `/etc/passwd` if `$HOME` is not found. For
Windows, the home directory is now obtained from `%USERPROFILE%` instead
of `%HOMEPATH%` as USERPROFILE also contains the drive letter and the
Windows bin has been given a better fallback which includes the system
drive.